changes as discussed in #32
