+++ src/cosmospy/_wallet.py
@@ -6,7 +6,7 @@ import hdwallets
 import mnemonic
 
 from cosmospy import BIP32DerivationError
-from cosmospy._typing import Wallet
+from cosmospy.typing import Wallet
 
 DEFAULT_DERIVATION_PATH = "m/44'/118'/0'/0/0"
 DEFAULT_BECH32_HRP = "cosmos"
