+++ src/cosmospy/typing.py
@@ -1,0 +1,14 @@
+import sys
+
+if sys.version_info < (3, 8):
+    from typing_extensions import TypedDict
+else:
+    from typing import TypedDict
+
+
+class Wallet(TypedDict):
+    seed: str
+    derivation_path: str
+    private_key: bytes
+    public_key: bytes
+    address: str
