+++ tests/test_wallet.py
@@ -3,7 +3,7 @@ from unittest import mock
 from ward import test
 
 from cosmospy import generate_wallet, privkey_to_address, privkey_to_pubkey, seed_to_privkey
-from cosmospy._typing import Wallet
+from cosmospy.typing import Wallet
 
 test_vector: Wallet = {
     "seed": "todo",
